feat: Switch to using uv for dependency management
Switch from using pip to using uv for dependency management. Not
only is the uv tool significantly faster, it creates lock files
by default for all of the project's declared dependencies which
aides in the endeavor to have reproducability for the charm.

This will modify dependency management workflows and will slightly
increase the time for charmcraft pack due to the installation of
the uv snap as a build time dependency. The trade-off for the
faster dependency resolver (for devs) and the better dependency
management is worth it.

Included in this change is migrating to use the tox-uv plugin,
which provides the uv-venv-lock-runner environments runner
for honoring the contents of the uv.lock file.

A contributing guide is included in order to help new users
unfamiliar with setting up the uv dependencies and setting up
a juju environment.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>